Fix some warning on Windows systems.

+++ usrsctplib/user_socket.c
@@ -3310,7 +3310,7 @@ usrsctp_conninput(void *addr, const void *buffer, size_t length, uint8_t ecn_bit
 	struct mbuf *m, *mm;
 	struct sctphdr *sh;
 	struct sctp_chunkhdr *ch;
-	size_t remaining;
+	int remaining;
 
 	SCTP_STAT_INCR(sctps_recvpackets);
 	SCTP_STAT_INCR_COUNTER64(sctps_inpackets);
@@ -3332,9 +3332,9 @@ usrsctp_conninput(void *addr, const void *buffer, size_t length, uint8_t ecn_bit
 	/* Set the lengths fields of the mbuf chain.
 	 * This is expected by m_copyback().
 	 */
-	remaining = length;
+	remaining = (int)length;
 	for (mm = m; mm != NULL; mm = mm->m_next) {
-		mm->m_len = min(M_SIZE(mm), remaining);
+		mm->m_len = min((int)M_SIZE(mm), remaining);
 		m->m_pkthdr.len += mm->m_len;
 		remaining -= mm->m_len;
 	}
